# fix: synchronize modification of shared calendar (#921)