--- a/pgjdbc/src/main/java/org/postgresql/jdbc/PgPreparedStatement.java
+++ b/pgjdbc/src/main/java/org/postgresql/jdbc/PgPreparedStatement.java
@@ -82,8 +82,6 @@ class PgPreparedStatement extends PgStatement implements PreparedStatement {
    */
   protected boolean outParmBeforeFunc = false;
 
-  private TimeZone defaultTimeZone;
-
   PgPreparedStatement(PgConnection connection, String sql, int rsType, int rsConcurrency,
       int rsHoldability) throws SQLException {
     this(connection, connection.borrowQuery(sql), rsType, rsConcurrency, rsHoldability);
@@ -170,7 +168,7 @@ class PgPreparedStatement extends PgStatement implements PreparedStatement {
 
       return (result != null && result.getResultSet() != null);
     } finally {
-      defaultTimeZone = null;
+      connection.getTimestampUtils().clearDefaultTimeZone();
     }
   }
 
@@ -590,7 +588,7 @@ class PgPreparedStatement extends PgStatement implements PreparedStatement {
             break;
             //#endif
           } else {
-            tmpd = connection.getTimestampUtils().toDate(getDefaultCalendar(), in.toString());
+            tmpd = connection.getTimestampUtils().toDate(null, in.toString());
           }
           setDate(parameterIndex, tmpd);
         }
@@ -608,7 +606,7 @@ class PgPreparedStatement extends PgStatement implements PreparedStatement {
             break;
             //#endif
           } else {
-            tmpt = connection.getTimestampUtils().toTime(getDefaultCalendar(), in.toString());
+            tmpt = connection.getTimestampUtils().toTime(null, in.toString());
           }
           setTime(parameterIndex, tmpt);
         }
@@ -628,7 +626,7 @@ class PgPreparedStatement extends PgStatement implements PreparedStatement {
             break;
             //#endif
           } else {
-            tmpts = connection.getTimestampUtils().toTimestamp(getDefaultCalendar(), in.toString());
+            tmpts = connection.getTimestampUtils().toTimestamp(null, in.toString());
           }
           setTimestamp(parameterIndex, tmpts);
         }
@@ -1289,9 +1287,6 @@ class PgPreparedStatement extends PgStatement implements PreparedStatement {
     // 2005-01-01 00:00:00+03
     // (1 row)
 
-    if (cal == null) {
-      cal = getDefaultCalendar();
-    }
     bindString(i, connection.getTimestampUtils().toString(cal, d), Oid.UNSPECIFIED);
   }
 
@@ -1316,9 +1311,6 @@ class PgPreparedStatement extends PgStatement implements PreparedStatement {
       }
     }
 
-    if (cal == null) {
-      cal = getDefaultCalendar();
-    }
     bindString(i, connection.getTimestampUtils().toString(cal, t), oid);
   }
 
@@ -1371,9 +1363,6 @@ class PgPreparedStatement extends PgStatement implements PreparedStatement {
         oid = Oid.TIMESTAMPTZ;
         cal = pgTimestamp.getCalendar();
       }
-    }
-    if (cal == null) {
-      cal = getDefaultCalendar();
     }
     bindString(i, connection.getTimestampUtils().toString(cal, t), oid);
   }
@@ -1562,20 +1551,8 @@ class PgPreparedStatement extends PgStatement implements PreparedStatement {
       }
       return super.executeBatch();
     } finally {
-      defaultTimeZone = null;
+      connection.getTimestampUtils().clearDefaultTimeZone();
     }
-  }
-
-  private Calendar getDefaultCalendar() {
-    TimestampUtils timestampUtils = connection.getTimestampUtils();
-    if (timestampUtils.hasFastDefaultTimeZone()) {
-      return timestampUtils.getSharedCalendar(null);
-    }
-    Calendar sharedCalendar = timestampUtils.getSharedCalendar(defaultTimeZone);
-    if (defaultTimeZone == null) {
-      defaultTimeZone = sharedCalendar.getTimeZone();
-    }
-    return sharedCalendar;
   }
 
   public ParameterMetaData getParameterMetaData() throws SQLException {