--- a/pgjdbc/src/main/java/org/postgresql/jdbc/PgResultSet.java
+++ b/pgjdbc/src/main/java/org/postgresql/jdbc/PgResultSet.java
@@ -92,7 +92,6 @@ public class PgResultSet implements ResultSet, org.postgresql.PGRefCursorResultS
   private final int resultsettype;
   private final int resultsetconcurrency;
   private int fetchdirection = ResultSet.FETCH_UNKNOWN;
-  private TimeZone defaultTimeZone;
   protected final BaseConnection connection; // the connection we belong to
   protected final BaseStatement statement; // the statement we belong to
   protected final Field[] fields; // Field metadata for this resultset.
@@ -471,13 +470,10 @@ public class PgResultSet implements ResultSet, org.postgresql.PGRefCursorResultS
       return null;
     }
 
-    if (cal == null) {
-      cal = getDefaultCalendar();
-    }
     if (isBinary(i)) {
       int col = i - 1;
       int oid = fields[col].getOID();
-      TimeZone tz = cal.getTimeZone();
+      TimeZone tz = (cal != null)? cal.getTimeZone() : null;
       if (oid == Oid.DATE) {
         return connection.getTimestampUtils().toDateBin(tz, this_row[col]);
       } else if (oid == Oid.TIMESTAMP || oid == Oid.TIMESTAMPTZ) {
@@ -505,13 +501,10 @@ public class PgResultSet implements ResultSet, org.postgresql.PGRefCursorResultS
       return null;
     }
 
-    if (cal == null) {
-      cal = getDefaultCalendar();
-    }
     if (isBinary(i)) {
       int col = i - 1;
       int oid = fields[col].getOID();
-      TimeZone tz = cal.getTimeZone();
+      TimeZone tz = (cal != null)? cal.getTimeZone() : null;
       if (oid == Oid.TIME || oid == Oid.TIMETZ) {
         return connection.getTimestampUtils().toTimeBin(tz, this_row[col]);
       } else if (oid == Oid.TIMESTAMP || oid == Oid.TIMESTAMPTZ) {
@@ -565,15 +558,12 @@ public class PgResultSet implements ResultSet, org.postgresql.PGRefCursorResultS
       return null;
     }
 
-    if (cal == null) {
-      cal = getDefaultCalendar();
-    }
     int col = i - 1;
     int oid = fields[col].getOID();
     if (isBinary(i)) {
       if (oid == Oid.TIMESTAMPTZ || oid == Oid.TIMESTAMP) {
         boolean hasTimeZone = oid == Oid.TIMESTAMPTZ;
-        TimeZone tz = cal.getTimeZone();
+        TimeZone tz = (cal != null)? cal.getTimeZone() : null;
         return connection.getTimestampUtils().toTimestampBin(tz, this_row[col], hasTimeZone);
       } else {
         // JDBC spec says getTimestamp of Time and Date must be supported
@@ -619,8 +609,7 @@ public class PgResultSet implements ResultSet, org.postgresql.PGRefCursorResultS
               PSQLState.DATA_TYPE_MISMATCH);
     }
     if (isBinary(i)) {
-      TimeZone timeZone = getDefaultCalendar().getTimeZone();
-      return connection.getTimestampUtils().toLocalDateTimeBin(timeZone, this_row[col]);
+      return connection.getTimestampUtils().toLocalDateTimeBin(null, this_row[col]);
     }
 
     String string = getString(i);
@@ -1713,20 +1702,20 @@ public class PgResultSet implements ResultSet, org.postgresql.PGRefCursorResultS
             rowBuffer[columnIndex] = connection
                 .encodeString(
                     connection.getTimestampUtils().toString(
-                        getDefaultCalendar(), (Date) valueObject));
+                        null, (Date) valueObject));
             break;
 
           case Types.TIME:
             rowBuffer[columnIndex] = connection
                 .encodeString(
                     connection.getTimestampUtils().toString(
-                        getDefaultCalendar(), (Time) valueObject));
+                        null, (Time) valueObject));
             break;
 
           case Types.TIMESTAMP:
             rowBuffer[columnIndex] = connection.encodeString(
                 connection.getTimestampUtils().toString(
-                    getDefaultCalendar(), (Timestamp) valueObject));
+                    null, (Timestamp) valueObject));
             break;
 
           case Types.NULL:
@@ -3253,7 +3242,7 @@ public class PgResultSet implements ResultSet, org.postgresql.PGRefCursorResultS
       //#endif
       ) {
         Timestamp timestampValue = getTimestamp(columnIndex);
-        Calendar calendar = Calendar.getInstance(getDefaultCalendar().getTimeZone());
+        Calendar calendar = Calendar.getInstance(connection.getTimestampUtils().getDefaultCalendarTimeZone());
         calendar.setTimeInMillis(timestampValue.getTime());
         return type.cast(calendar);
       } else {
@@ -3637,15 +3626,4 @@ public class PgResultSet implements ResultSet, org.postgresql.PGRefCursorResultS
     throw new SQLException("Cannot unwrap to " + iface.getName());
   }
 
-  private Calendar getDefaultCalendar() {
-    TimestampUtils timestampUtils = connection.getTimestampUtils();
-    if (timestampUtils.hasFastDefaultTimeZone()) {
-      return timestampUtils.getSharedCalendar(null);
-    }
-    Calendar sharedCalendar = timestampUtils.getSharedCalendar(defaultTimeZone);
-    if (defaultTimeZone == null) {
-      defaultTimeZone = sharedCalendar.getTimeZone();
-    }
-    return sharedCalendar;
-  }
 }