Calendar is not thread-safe, so ensuring all entry points that modify shared instances of `org.postgresql.jdbc.TimestampUtils.calendarWithUserTz` are properly synchronized.

fixes #921

---

Simply ensuring the following methods:

- `TimestampUtils.convertToDate(long,TimeZone)`
- `TimestampUtils.convertToTime(long,TimeZone)`
- `TimestampUtils.guessTimestamp(long,TimeZone)`

are synchronized, as they may modify the shared calendar `calendarWithUserTz` via `java.util.Calendar.set()` methods.

I have added a failing unit test to `org.postgresql.test.jdbc2.DateTest`, which outputs something like

```
   testDateThreadSafety(org.postgresql.test.jdbc2.DateTest)  Time elapsed: 0.301 sec  <<< FAILURE!
   java.lang.AssertionError: expected:<2017> but was:<7777>
   	at org.junit.Assert.fail(Assert.java:88)
   	at org.junit.Assert.failNotEquals(Assert.java:834)
   	at org.junit.Assert.assertEquals(Assert.java:645)
   	at org.junit.Assert.assertEquals(Assert.java:631)
   	at org.postgresql.test.jdbc2.DateTest.testDateThreadSafety(DateTest.java:216)
```

and confirmed it passes under JDK 8, 7, and 6 ("BEFORE", "AFTER", "JDK7", and "JDK6", respectively, below).

```
--- [BEFORE: mvn clean package] ---

johnsontm@dellxps139360:~/pgjdbc/git/pgjdbc$ mvn clean package
[INFO] Scanning for projects...
[INFO] ------------------------------------------------------------------------
[INFO] Reactor Build Order:
[INFO] 
[INFO] PostgreSQL JDBC Driver - JDBC 4.2
[INFO] PostgreSQL JDBC Driver - benchmarks
[INFO] PostgreSQL JDBC Driver aggregate
[INFO]                                                                         
[INFO] ------------------------------------------------------------------------
[INFO] Building PostgreSQL JDBC Driver - JDBC 4.2 42.1.5-SNAPSHOT
[INFO] ------------------------------------------------------------------------
[INFO] 
[INFO] --- maven-clean-plugin:2.5:clean (default-clean) @ postgresql ---
[INFO] 
[INFO] --- maven-enforcer-plugin:1.3.1:enforce (enforce-java) @ postgresql ---
[INFO] 
[INFO] --- build-helper-maven-plugin:1.5:parse-version (parse-version) @ postgresql ---
[INFO] 
[INFO] --- properties-maven-plugin:1.0-alpha-2:read-project-properties (default) @ postgresql ---
[INFO] 
[INFO] --- jcp:6.0.1:preprocess (preprocessSources) @ postgresql ---
[INFO] Added MAVEN property mvn.project.name=PostgreSQL JDBC Driver - JDBC 4.2
[INFO] Added MAVEN property mvn.project.version=42.1.5-SNAPSHOT
[INFO] Added MAVEN property mvn.project.url=https://github.com/pgjdbc/pgjdbc
[INFO] Added MAVEN property mvn.project.packaging=bundle
[INFO] Added MAVEN property mvn.project.modelversion=4.0.0
[INFO] Added MAVEN property mvn.project.inceptionyear=
[INFO] Added MAVEN property mvn.project.id=org.postgresql:postgresql:bundle:42.1.5-SNAPSHOT
[INFO] Added MAVEN property mvn.project.groupid=org.postgresql
[INFO] Added MAVEN property mvn.project.description=Java JDBC 4.2 (JRE 8+) driver for PostgreSQL database
[INFO] Added MAVEN property mvn.project.artifact.id=org.postgresql:postgresql:bundle:42.1.5-SNAPSHOT
[INFO] Added MAVEN property mvn.project.artifact.artifactid=postgresql
[INFO] Added MAVEN property mvn.project.artifact.baseversion=42.1.5-SNAPSHOT
[INFO] Added MAVEN property mvn.project.artifact.dependencyconflictid=org.postgresql:postgresql:bundle
[INFO] Added MAVEN property mvn.project.artifact.downloadurl=
[INFO] Added MAVEN property mvn.project.artifact.groupid=org.postgresql
[INFO] Added MAVEN property mvn.project.artifact.scope=
[INFO] Added MAVEN property mvn.project.artifact.type=bundle
[INFO] Added MAVEN property mvn.project.artifact.version=42.1.5-SNAPSHOT
[INFO] Added MAVEN property mvn.project.build.directory=/home/johnsontm/pgjdbc/git/pgjdbc/pgjdbc/target
[INFO] Added MAVEN property mvn.project.build.defaultgoal=
[INFO] Added MAVEN property mvn.project.build.outputdirectory=/home/johnsontm/pgjdbc/git/pgjdbc/pgjdbc/target/classes
[INFO] Added MAVEN property mvn.project.build.scriptsourcedirectory=/home/johnsontm/pgjdbc/git/pgjdbc/pgjdbc/src/main/scripts
[INFO] Added MAVEN property mvn.project.build.sourcedirectory=/home/johnsontm/pgjdbc/git/pgjdbc/pgjdbc/src/main/java
[INFO] Added MAVEN property mvn.project.build.testoutputdirectory=/home/johnsontm/pgjdbc/git/pgjdbc/pgjdbc/target/test-classes
[INFO] Added MAVEN property mvn.project.build.testsourcedirectory=/home/johnsontm/pgjdbc/git/pgjdbc/pgjdbc/src/test/java
[INFO] Added MAVEN property mvn.project.organization.name=PostgreSQL Global Development Group
[INFO] Added MAVEN property mvn.project.organization.url=https://jdbc.postgresql.org/
[INFO] Added MAVEN property mvn.project.activeprofiles=
[INFO] Added MAVEN property mvn.project.property.postgresql.enforce.jdk.version=1.8
[INFO] Added MAVEN property mvn.project.property.postgresql.preprocessed.test.sources.directory=/home/johnsontm/pgjdbc/git/pgjdbc/pgjdbc/target/gen-test-src
[INFO] Added MAVEN property mvn.project.property.loggerlevel=OFF
[INFO] Added MAVEN property mvn.project.property.javac.target=1.8
[INFO] Added MAVEN property mvn.project.property.postgresql.driver.fullversion=PostgreSQL JDBC4.2
[INFO] Added MAVEN property mvn.project.property.jdbc.specification.version=4.2
[INFO] Added MAVEN property mvn.project.property.unzip.jdk.ant.build.xml=/home/johnsontm/pgjdbc/git/pgjdbc/pgjdbc/target/unzip-jdk-ant.xml
[INFO] Added MAVEN property mvn.project.property.loggerfile=target/pgjdbc-tests.log
[INFO] Added MAVEN property mvn.project.property.parsedversion.minorversion=1
[INFO] Added MAVEN property mvn.project.property.project.build.sourceencoding=UTF-8
[INFO] Added MAVEN property mvn.project.property.build.properties.relative.path=/home/johnsontm/pgjdbc/git/pgjdbc/pgjdbc/..
[INFO] Added MAVEN property mvn.project.property.privilegedpassword=***** [hidden because may contain private info]
[INFO] Added MAVEN property mvn.project.property.waffleenabled=true
[INFO] Added MAVEN property mvn.project.property.skip.assembly=false
[INFO] Added MAVEN property mvn.project.property.protocolversion=0
[INFO] Added MAVEN property mvn.project.property.postgresql.preprocessed.sources.directory=/home/johnsontm/pgjdbc/git/pgjdbc/pgjdbc/target/gen-src
[INFO] Added MAVEN property mvn.project.property.osgienabled=true
[INFO] Added MAVEN property mvn.project.property.preparethreshold=5
[INFO] Added MAVEN property mvn.project.property.password=***** [hidden because may contain private info]
[INFO] Added MAVEN property mvn.project.property.jdbc.specification.version.nodot=42
[INFO] Added MAVEN property mvn.project.property.username=test
[INFO] Added MAVEN property mvn.project.property.parsedversion.buildnumber=0
[INFO] Added MAVEN property mvn.project.property.server=localhost
[INFO] Added MAVEN property mvn.project.property.port=5432
[INFO] Added MAVEN property mvn.project.property.template.default.pg.port=5432
[INFO] Added MAVEN property mvn.project.property.parsedversion.qualifier=SNAPSHOT
[INFO] Added MAVEN property mvn.project.property.waffle-jna.version=1.7.5
[INFO] Added MAVEN property mvn.project.property.skip.unzip-jdk-src=true
[INFO] Added MAVEN property mvn.project.property.database=test
[INFO] Added MAVEN property mvn.project.property.parsedversion.incrementalversion=5
[INFO] Added MAVEN property mvn.project.property.sspiusername=testsspi
[INFO] Added MAVEN property mvn.project.property.privilegeduser=postgres
[INFO] Added MAVEN property mvn.project.property.checkstyle.version=7.8.2
[INFO] Added MAVEN property mvn.project.property.parsedversion.majorversion=42
[INFO] Added MAVEN property mvn.project.property.parsedversion.osgiversion=42.1.5.SNAPSHOT
[INFO] Added MAVEN property mvn.project.property.postgresql.jdbc.spec=JDBC4.2
[INFO] Added MAVEN property mvn.project.property.argline=-Xmx512m
[INFO] Preprocess sources : /home/johnsontm/pgjdbc/git/pgjdbc/pgjdbc/src/main/java
[INFO] Preprocess destination : /home/johnsontm/pgjdbc/git/pgjdbc/pgjdbc/target/gen-src
[INFO] -----------------------------------------------------------------
[INFO] Completed, preprocessed 208 files, copied 16 files, elapsed time 355ms
[INFO] A Compile source root has been removed from the root list [/home/johnsontm/pgjdbc/git/pgjdbc/pgjdbc/src/main/java]
[INFO] The New compile source root has been added into the list [/home/johnsontm/pgjdbc/git/pgjdbc/pgjdbc/target/gen-src]
[INFO] 
[INFO] --- maven-resources-plugin:2.6:resources (default-resources) @ postgresql ---
[INFO] Using 'UTF-8' encoding to copy filtered resources.
[INFO] Copying 17 resources
[INFO] 
[INFO] --- maven-compiler-plugin:3.3:compile (default-compile) @ postgresql ---
[INFO] Changes detected - recompiling the module!
[INFO] Compiling 208 source files to /home/johnsontm/pgjdbc/git/pgjdbc/pgjdbc/target/classes
[INFO] /home/johnsontm/pgjdbc/git/pgjdbc/pgjdbc/target/gen-src/org/postgresql/jdbc3/Jdbc3PoolingDataSource.java: Some input files use or override a deprecated API.
[INFO] /home/johnsontm/pgjdbc/git/pgjdbc/pgjdbc/target/gen-src/org/postgresql/jdbc3/Jdbc3PoolingDataSource.java: Recompile with -Xlint:deprecation for details.
[INFO] /home/johnsontm/pgjdbc/git/pgjdbc/pgjdbc/target/gen-src/org/postgresql/util/LruCache.java: Some input files use unchecked or unsafe operations.
[INFO] /home/johnsontm/pgjdbc/git/pgjdbc/pgjdbc/target/gen-src/org/postgresql/util/LruCache.java: Recompile with -Xlint:unchecked for details.
[INFO] 
[INFO] --- jcp:6.0.1:preprocess (preprocessTestSources) @ postgresql ---
[INFO] Added MAVEN property mvn.project.name=PostgreSQL JDBC Driver - JDBC 4.2
[INFO] Added MAVEN property mvn.project.version=42.1.5-SNAPSHOT
[INFO] Added MAVEN property mvn.project.url=https://github.com/pgjdbc/pgjdbc
[INFO] Added MAVEN property mvn.project.packaging=bundle
[INFO] Added MAVEN property mvn.project.modelversion=4.0.0
[INFO] Added MAVEN property mvn.project.inceptionyear=
[INFO] Added MAVEN property mvn.project.id=org.postgresql:postgresql:bundle:42.1.5-SNAPSHOT
[INFO] Added MAVEN property mvn.project.groupid=org.postgresql
[INFO] Added MAVEN property mvn.project.description=Java JDBC 4.2 (JRE 8+) driver for PostgreSQL database
[INFO] Added MAVEN property mvn.project.artifact.id=org.postgresql:postgresql:bundle:42.1.5-SNAPSHOT
[INFO] Added MAVEN property mvn.project.artifact.artifactid=postgresql
[INFO] Added MAVEN property mvn.project.artifact.baseversion=42.1.5-SNAPSHOT
[INFO] Added MAVEN property mvn.project.artifact.dependencyconflictid=org.postgresql:postgresql:bundle
[INFO] Added MAVEN property mvn.project.artifact.downloadurl=
[INFO] Added MAVEN property mvn.project.artifact.groupid=org.postgresql
[INFO] Added MAVEN property mvn.project.artifact.scope=
[INFO] Added MAVEN property mvn.project.artifact.type=bundle
[INFO] Added MAVEN property mvn.project.artifact.version=42.1.5-SNAPSHOT
[INFO] Added MAVEN property mvn.project.build.directory=/home/johnsontm/pgjdbc/git/pgjdbc/pgjdbc/target
[INFO] Added MAVEN property mvn.project.build.defaultgoal=
[INFO] Added MAVEN property mvn.project.build.outputdirectory=/home/johnsontm/pgjdbc/git/pgjdbc/pgjdbc/target/classes
[INFO] Added MAVEN property mvn.project.build.scriptsourcedirectory=/home/johnsontm/pgjdbc/git/pgjdbc/pgjdbc/src/main/scripts
[INFO] Added MAVEN property mvn.project.build.sourcedirectory=/home/johnsontm/pgjdbc/git/pgjdbc/pgjdbc/src/main/java
[INFO] Added MAVEN property mvn.project.build.testoutputdirectory=/home/johnsontm/pgjdbc/git/pgjdbc/pgjdbc/target/test-classes
[INFO] Added MAVEN property mvn.project.build.testsourcedirectory=/home/johnsontm/pgjdbc/git/pgjdbc/pgjdbc/src/test/java
[INFO] Added MAVEN property mvn.project.organization.name=PostgreSQL Global Development Group
[INFO] Added MAVEN property mvn.project.organization.url=https://jdbc.postgresql.org/
[INFO] Added MAVEN property mvn.project.activeprofiles=
[INFO] Added MAVEN property mvn.project.property.postgresql.enforce.jdk.version=1.8
[INFO] Added MAVEN property mvn.project.property.postgresql.preprocessed.test.sources.directory=/home/johnsontm/pgjdbc/git/pgjdbc/pgjdbc/target/gen-test-src
[INFO] Added MAVEN property mvn.project.property.loggerlevel=OFF
[INFO] Added MAVEN property mvn.project.property.javac.target=1.8
[INFO] Added MAVEN property mvn.project.property.postgresql.driver.fullversion=PostgreSQL JDBC4.2
[INFO] Added MAVEN property mvn.project.property.jdbc.specification.version=4.2
[INFO] Added MAVEN property mvn.project.property.unzip.jdk.ant.build.xml=/home/johnsontm/pgjdbc/git/pgjdbc/pgjdbc/target/unzip-jdk-ant.xml
[INFO] Added MAVEN property mvn.project.property.loggerfile=target/pgjdbc-tests.log
[INFO] Added MAVEN property mvn.project.property.parsedversion.minorversion=1
[INFO] Added MAVEN property mvn.project.property.project.build.sourceencoding=UTF-8
[INFO] Added MAVEN property mvn.project.property.build.properties.relative.path=/home/johnsontm/pgjdbc/git/pgjdbc/pgjdbc/..
[INFO] Added MAVEN property mvn.project.property.privilegedpassword=***** [hidden because may contain private info]
[INFO] Added MAVEN property mvn.project.property.waffleenabled=true
[INFO] Added MAVEN property mvn.project.property.skip.assembly=false
[INFO] Added MAVEN property mvn.project.property.protocolversion=0
[INFO] Added MAVEN property mvn.project.property.postgresql.preprocessed.sources.directory=/home/johnsontm/pgjdbc/git/pgjdbc/pgjdbc/target/gen-src
[INFO] Added MAVEN property mvn.project.property.osgienabled=true
[INFO] Added MAVEN property mvn.project.property.preparethreshold=5
[INFO] Added MAVEN property mvn.project.property.password=***** [hidden because may contain private info]
[INFO] Added MAVEN property mvn.project.property.jdbc.specification.version.nodot=42
[INFO] Added MAVEN property mvn.project.property.username=test
[INFO] Added MAVEN property mvn.project.property.parsedversion.buildnumber=0
[INFO] Added MAVEN property mvn.project.property.server=localhost
[INFO] Added MAVEN property mvn.project.property.port=5432
[INFO] Added MAVEN property mvn.project.property.template.default.pg.port=5432
[INFO] Added MAVEN property mvn.project.property.parsedversion.qualifier=SNAPSHOT
[INFO] Added MAVEN property mvn.project.property.waffle-jna.version=1.7.5
[INFO] Added MAVEN property mvn.project.property.skip.unzip-jdk-src=true
[INFO] Added MAVEN property mvn.project.property.database=test
[INFO] Added MAVEN property mvn.project.property.parsedversion.incrementalversion=5
[INFO] Added MAVEN property mvn.project.property.sspiusername=testsspi
[INFO] Added MAVEN property mvn.project.property.privilegeduser=postgres
[INFO] Added MAVEN property mvn.project.property.checkstyle.version=7.8.2
[INFO] Added MAVEN property mvn.project.property.parsedversion.majorversion=42
[INFO] Added MAVEN property mvn.project.property.parsedversion.osgiversion=42.1.5.SNAPSHOT
[INFO] Added MAVEN property mvn.project.property.postgresql.jdbc.spec=JDBC4.2
[INFO] Added MAVEN property mvn.project.property.argline=-Xmx512m
[INFO] Preprocess sources : /home/johnsontm/pgjdbc/git/pgjdbc/pgjdbc/src/test/java
[INFO] Preprocess destination : /home/johnsontm/pgjdbc/git/pgjdbc/pgjdbc/target/gen-test-src
[INFO] -----------------------------------------------------------------
[INFO] Completed, preprocessed 145 files, copied 2 files, elapsed time 93ms
[INFO] A Compile source root has been removed from the root list [/home/johnsontm/pgjdbc/git/pgjdbc/pgjdbc/src/test/java]
[INFO] The New compile source root has been added into the list [/home/johnsontm/pgjdbc/git/pgjdbc/pgjdbc/target/gen-test-src]
[INFO] 
[INFO] --- maven-resources-plugin:2.6:testResources (default-testResources) @ postgresql ---
[INFO] Using 'UTF-8' encoding to copy filtered resources.
[INFO] Copying 1 resource
[INFO] 
[INFO] --- maven-compiler-plugin:3.3:testCompile (default-testCompile) @ postgresql ---
[INFO] Changes detected - recompiling the module!
[INFO] Compiling 145 source files to /home/johnsontm/pgjdbc/git/pgjdbc/pgjdbc/target/test-classes
[INFO] /home/johnsontm/pgjdbc/git/pgjdbc/pgjdbc/target/gen-test-src/org/postgresql/test/jdbc2/optional/CaseOptimiserDataSourceTest.java: Some input files use or override a deprecated API.
[INFO] /home/johnsontm/pgjdbc/git/pgjdbc/pgjdbc/target/gen-test-src/org/postgresql/test/jdbc2/optional/CaseOptimiserDataSourceTest.java: Recompile with -Xlint:deprecation for details.
[INFO] /home/johnsontm/pgjdbc/git/pgjdbc/pgjdbc/target/gen-test-src/org/postgresql/test/hostchooser/MultiHostsConnectionTest.java: Some input files use unchecked or unsafe operations.
[INFO] /home/johnsontm/pgjdbc/git/pgjdbc/pgjdbc/target/gen-test-src/org/postgresql/test/hostchooser/MultiHostsConnectionTest.java: Recompile with -Xlint:unchecked for details.
[INFO] 
[INFO] --- maven-surefire-plugin:2.18.1:test (default-test) @ postgresql ---
[INFO] Surefire report directory: /home/johnsontm/pgjdbc/git/pgjdbc/pgjdbc/target/surefire-reports

-------------------------------------------------------
 T E S T S
-------------------------------------------------------
Skipping SingleCertSocketFactoryTests. To enable set the property testsinglecertfactory=true in the ssltest.properties file.
Skipping ssloff8.
Skipping sslhostnossl8.
Skipping ssloff9.
Skipping sslhostnossl9.
Skipping sslhostgh8.
Skipping sslhostgh9.
Skipping sslhostbh8.
Skipping sslhostbh9.
Skipping sslhostsslgh8.
Skipping sslhostsslgh9.
Skipping sslhostsslbh8.
Skipping sslhostsslbh9.
Skipping sslhostsslcertgh8.
Skipping sslhostsslcertgh9.
Skipping sslhostsslcertbh8.
Skipping sslhostsslcertbh9.
Skipping sslcertgh8.
Skipping sslcertgh9.
Skipping sslcertbh8.
Skipping sslcertbh9.
Running org.postgresql.test.extensions.ExtensionsTestSuite
Tests run: 5, Failures: 0, Errors: 0, Skipped: 5, Time elapsed: 0.079 sec - in org.postgresql.test.extensions.ExtensionsTestSuite
Running org.postgresql.test.osgi.OsgiTestSuite
Tests run: 8, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.001 sec - in org.postgresql.test.osgi.OsgiTestSuite
Running org.postgresql.test.ssl.SingleCertValidatingFactoryTestSuite
Skipping SingleCertSocketFactoryTests. To enable set the property testsinglecertfactory=true in the ssltest.properties file.
Tests run: 0, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.001 sec - in org.postgresql.test.ssl.SingleCertValidatingFactoryTestSuite
Running org.postgresql.test.ssl.SslTestSuite
Skipping ssloff8.
Skipping sslhostnossl8.
Skipping ssloff9.
Skipping sslhostnossl9.
Skipping sslhostgh8.
Skipping sslhostgh9.
Skipping sslhostbh8.
Skipping sslhostbh9.
Skipping sslhostsslgh8.
Skipping sslhostsslgh9.
Skipping sslhostsslbh8.
Skipping sslhostsslbh9.
Skipping sslhostsslcertgh8.
Skipping sslhostsslcertgh9.
Skipping sslhostsslcertbh8.
Skipping sslhostsslcertbh9.
Skipping sslcertgh8.
Skipping sslcertgh9.
Skipping sslcertbh8.
Skipping sslcertbh9.
Tests run: 0, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0 sec - in org.postgresql.test.ssl.SslTestSuite
Running org.postgresql.test.jdbc2.Jdbc2TestSuite
Using seed = 1505089562717 for StrangeInputStream. Set -DStrangeInputStream.seed=1505089562717 to reproduce the test
Using seed = 1505089562949 for StrangeInputStream. Set -DStrangeInputStream.seed=1505089562949 to reproduce the test
Using seed = 1505089563195 for StrangeInputStream. Set -DStrangeInputStream.seed=1505089563195 to reproduce the test
Using seed = 1505089563457 for StrangeInputStream. Set -DStrangeInputStream.seed=1505089563457 to reproduce the test
Using seed = 1505089563678 for StrangeInputStream. Set -DStrangeInputStream.seed=1505089563678 to reproduce the test
Using seed = 1505089563915 for StrangeInputStream. Set -DStrangeInputStream.seed=1505089563915 to reproduce the test
Using seed = 1505089564124 for StrangeInputStream. Set -DStrangeInputStream.seed=1505089564124 to reproduce the test
…