@sehrope 2017-09-11
Why is this necessary? `TimestampUtils` is local to each connection and your test is accessing the same connection from two different threads with no external synchronization. That type of non-synchronized access is likely to encounter other issues as well. Connection should not be used concurrently for query execution by multiple threads.