@vlsi 2017-09-11
>That type of non-synchronized access is likely to encounter other issues as well.

I agree.

>Connection should not be used concurrently for query execution by multiple threads.

@sehrope , the thing is pgjdbc [claimed](https://jdbc.postgresql.org/documentation/head/thread.html) to be fully thread-safe:

> doc: The PostgreSQL™ JDBC driver is thread safe

See the suggested change for documentation: https://github.com/pgjdbc/pgjdbc/pull/928