@sehrope 2017-09-11
@vlsi Wow I didn't know the docs said that! Yes that's blatantly false and should be removed from the docs.