@vlsi 2017-09-11
>Wow I didn't know the docs said that! Yes that's blatantly false and should be removed from the docs.

On the other hand, threading issues (e.g. reusing a connection across threads) are quite common, so adding some safety still makes sense.