@sehrope 2017-09-11
> On the other hand, threading issues (e.g. reusing a connection across threads) are quite common, so adding some safety still makes sense.

Not if it's going to be for something that's internal or not required to be safe. Otherwise it's just added complexity and overhead.

Unless the entire usage of a Connection is thread safe there's no real gain in being partially thread safe anyway. If anything it's scarier as other unhandled edge cases would be encountered.

Better to be explicit that it's not intended to be used by multiple threads (outside of say `.cancel()`) and not deal with any of this.