@davecramer 2017-09-11
agreed, lets just remove the notion that there is any thread safety in the driver