@vlsi 2017-09-11
> lets just remove the notion that there is any thread safety in the driver

There's safety for `.cancel()`.
There's a patch for `.getWarnings()` safety