@davecramer 2017-09-11
Fair enough. But lets not encourage the notion that this can be used in any multi-threaded application.