@tordmjohnson 2017-09-11
OK. I completely understand nobody wants to go down the 'slipper slope' of trying to make ``Connection`` entirely-thread safe. I just thought this was a simple enough change, and consistent with already existing synchronization in place for `TimestampUtils` methods like `toTimestamp(Calendar,String)`, `toDate(Calendar,String)`, `toTime(Calendar,String)`, etc., which _do_ ensure the driver is thread-safe (with respect to ``Calendar``), at least when not using binary transfer.

If you're truly only going to change the documentation, then perhaps the existing synchronization in ``TimestampUtils`` should be removed instead, to ensure identical (mis)behavior regardless of whether using binary transfer or not.