@jorsol 2017-09-11
> I agree, however it turns out lots of uses depend on Statement.getWarnings thread safety. E.g. #927 

Fully agree, as I said, some methods require thread safety, and it's fine for these, but not necessarily for all.

> I think added synchronization would not matter much performance-wise. Nor it would matter support-wise: some of the methods have already been synchronized, so it could make sense to make them aligned.

I haven't made any performance measure of performance implications of synchronization, maybe new JVM works better and the "synchronization is expensive" no longer apply, so if this is the case, then it make sense to have a thread-safety driver.