@vlsi 2017-10-20
<img width="1241" alt="getsharedcalendar" src="https://user-images.githubusercontent.com/213894/31817484-d06ff3fe-b59c-11e7-86a8-d8b26b9fb927.png">


@tordmjohnson , would you please treat the other callers of `getSharedCalendar` as well?
`org.postgresql.jdbc.PgResultSet#getDefaultCalendar`
`org.postgresql.jdbc.PgPreparedStatement#getDefaultCalendar`