@tordmjohnson 2017-10-20
> Would you please treat PgResultSet#getDefaultCalendar and PgPreparedStatement#getDefaultCalendar?

I'd be happy to, @vlsi, but I'm not sure I understand exactly what you're asking for.

Are you simply asking that `PgResultSet#getDefaultCalendar()` and `PgPreparedStatement#getDefaultCalendar()` synchronize over the monitor for modifying `TimestampUtils#calendarWithUserTz` (the `TimestampUtils` instance itself), since they may indirectly set the time zone on the shared calendar? That is, something like:

```
   private Calendar getDefaultCalendar() {
      TimestampUtils timestampUtils = connection.getTimestampUtils();
      synchronized (timestampUtils) {
         ...
      }
      ...
   }
```