@vlsi 2017-10-20
@tordmjohnson , the idea is as follows: shared calendar should not be used outside of a synchronized context.

Things like the following should be simple. I bet just passing `null` should be enough (and remove the `if`)
```
    if (cal == null) {
      cal = getDefaultCalendar();
    }
    bindString(i, connection.getTimestampUtils().toString(cal, d), Oid.UNSPECIFIED);
```

org.postgresql.jdbc.PgResultSet#getDefaultCalendar is somewhat more involved.
However idea is the same: since `Calendar` is not thread-safe, there should be some synchronization