@vlsi 2017-10-20
A conclusion: `org.postgresql.jdbc.TimestampUtils#getSharedCalendar` should be private to avoid leakage of `Calendar` to prevent non-synchronized access to the calendar in question